+++ CHANGELOG.md
@@ -16,7 +16,7 @@
   * gpfs_fs_metadata_total_bytes to gpfs_fs_metadata_size_bytes
 * Removed metrics that can be calculated using other metrics
   * gpfs_fs_metadata_free_percent
-  * gpfs_fs_free_percent metrics
+  * gpfs_fs_free_percent
 * Remove nodename label from gpfs_perf_* metrics, replace with gpfs_perf_info metric
 * mmces and mmhealth status metrics will always have value 1, only the `status` label will change